+++ Magic/src/main/java/com/elmakers/mine/bukkit/world/generator/builtin/SequenceGenerator.java
@@ -37,10 +37,78 @@ public class SequenceGenerator extends BaseChunkGenerator {
         }
     }
 
+    @Override
+    public void generateCaves(@NotNull WorldInfo worldInfo, @NotNull Random random, int chunkX, int chunkZ, @NotNull ChunkData chunk) {
+        for (BaseChunkGenerator generator : generators) {
+            generator.generateCaves(worldInfo, random, chunkX, chunkZ, chunk);
+        }
+    }
+
+    @Override
     public void populate(WorldInfo worldInfo, Random random, int chunkX, int chunkZ, LimitedRegion region) {
         super.populate(worldInfo, random, chunkX, chunkZ, region);
         for (BaseChunkGenerator generator : generators) {
             generator.populate(worldInfo, random, chunkX, chunkZ, region);
         }
+    }
+
+    @Override
+    public boolean shouldGenerateNoise(@NotNull WorldInfo worldInfo, @NotNull Random random, int chunkX, int chunkZ) {
+        for (BaseChunkGenerator generator : generators) {
+            if (generator.shouldGenerateNoise(worldInfo, random, chunkX, chunkZ)) {
+                return true;
+            }
+        }
+        return false;
+    }
+
+    @Override
+    public boolean shouldGenerateSurface(@NotNull WorldInfo worldInfo, @NotNull Random random, int chunkX, int chunkZ) {
+        for (BaseChunkGenerator generator : generators) {
+            if (generator.shouldGenerateNoise(worldInfo, random, chunkX, chunkZ)) {
+                return true;
+            }
+        }
+        return false;
+    }
+
+    @Override
+    public boolean shouldGenerateCaves(@NotNull WorldInfo worldInfo, @NotNull Random random, int chunkX, int chunkZ) {
+        for (BaseChunkGenerator generator : generators) {
+            if (generator.shouldGenerateNoise(worldInfo, random, chunkX, chunkZ)) {
+                return true;
+            }
+        }
+        return false;
+    }
+
+    @Override
+    public boolean shouldGenerateDecorations(@NotNull WorldInfo worldInfo, @NotNull Random random, int chunkX, int chunkZ) {
+        for (BaseChunkGenerator generator : generators) {
+            if (generator.shouldGenerateNoise(worldInfo, random, chunkX, chunkZ)) {
+                return true;
+            }
+        }
+        return false;
+    }
+
+    @Override
+    public boolean shouldGenerateMobs(@NotNull WorldInfo worldInfo, @NotNull Random random, int chunkX, int chunkZ) {
+        for (BaseChunkGenerator generator : generators) {
+            if (generator.shouldGenerateNoise(worldInfo, random, chunkX, chunkZ)) {
+                return true;
+            }
+        }
+        return false;
+    }
+
+    @Override
+    public boolean shouldGenerateStructures(@NotNull WorldInfo worldInfo, @NotNull Random random, int chunkX, int chunkZ) {
+        for (BaseChunkGenerator generator : generators) {
+            if (generator.shouldGenerateNoise(worldInfo, random, chunkX, chunkZ)) {
+                return true;
+            }
+        }
+        return false;
     }
 }