+++ Magic/CHANGELOG.md
@@ -12,6 +12,7 @@
    - Sequence: Run multiple other generators in order
    - Random: Randomly choose one generator from a weighted probability map
    - Tower: Build a tower
+   - Vanilla: Use vanilla world generation
    - Schematic: Place a schematic in the world (see https://github.com/elBukkit/MagicPlugin/tree/main/Magic/src/examples/chaos)
  - Create block populators via `populators` configuration
    - These can use pre-existing populator classes, such as Replace, MagicChest, MagicBlock and Chest